+++ wire-android-sync-engine/zmessaging/src/main/scala/com/waz/model2/transport/responses/FetchSsoResponse.scala
@@ -9,7 +9,7 @@ case class SSOFound(sso: String) extends FetchSsoResponse
 
 object FetchSsoResponse {
 
-  implicit val fetchSsoResponseResponseDecoder: JsonDecoder[FetchSsoResponse] = new JsonDecoder[FetchSsoResponse] {
+  implicit val fetchSsoResponseDecoder: JsonDecoder[FetchSsoResponse] = new JsonDecoder[FetchSsoResponse] {
     override def apply(implicit js: JSONObject): FetchSsoResponse = {
       import JsonDecoder._
       val configUrl = decodeOptString('default_sso_code)
